gstnxt_backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx
pandas
openpyxl
xlsxwriter
python-dotenv
aiofiles
python-decouple
pytest
pytest-asyncio
